recommender/requirements.txt
google-cloud-bigquery == 3.11.4
pandas
db-dtypes
jupyterlab